[FIX] self_tests: Adjust expected values after pylint change
In a recent change on pylint-odoo [1], it was introduced a fix to
recognize a badly-formatted version in manifest, which now depends on
the Odoo version. When the version is provided and it's a numeric
version value (e.g. not "master") it's taken into account to determine
whether the version defined in the manifest is formatted correctly.

Since versions in tested modules are as follows:
- Two set to 8.0
- Four set to wrong values (e.g. 1.0)

When the version is numeric but it's not 8.0, there will be six errors,
otherwise there will be four. In these tests, the only case when it's
numeric but it's not 8.0, is when testing 7.0.

[1] Pylint change: https://github.com/OCA/pylint-odoo/commit/6b1f1e1bfaf